Unloose the Gordian Knot
* Split Code/Copy Lock/Lok.
